config.yml update + database schema rewrite
Added table to the auth method in config.yml

Database schema for the Account/AccountAuth/AccountAuthLocal
modules updated so that it has the full table definitions in
the modules. This is in preparation for a database schema
deployment script that will create the database according to
what is in the config.yml under the database definition and
the individual tables defined in the Perl modules.