Beginning of the map modules
Map locations will be identified either by a unique id or by
x,y,z and world co-ordinates. Tiles within the same world are
assumed to be on the same scale, but they don't have to be.
It allows us to separate inside a building from outside the
building with an entrance from one tile only (using an 'exit'
that links to a map id on a different world).