Fixed a bunch of accidental UTF-8 characters
A number of files have the UTF-8 'NO-BREAK SPACE' character
instead of a normal space. This is after a # character because
vim sometimes added it because alt-3 is a hash on a mac and
alt-space adds the no-break space.

This shows up as "#| " but deleting the | deletes both the |
and the space after because it's a double-byte character
0xC2 0xA0